Added compatibility to only delete finished schematic files and fixed one bug

+++ src/main/java/com/alpsbte/plotsystem/utils/ftp/FTPManager.java
@@ -91,17 +91,19 @@ public class FTPManager {
         return CompletableFuture.completedFuture(null);
     }
 
-    public static CompletableFuture<Void> deleteSchematics(String ftpURL, String schematicName) throws FileSystemException {
+    public static CompletableFuture<Void> deleteSchematics(String ftpURL, String schematicName, boolean onlyFinished) throws FileSystemException {
         try (StandardFileSystemManager fileManager = new StandardFileSystemManager()) {
             fileManager.init();
             FileObject remote, remoteSchematic;
 
-            remote = fileManager.resolveFile(ftpURL.replaceFirst("finishedSchematics/",""), fileOptions);
-            remoteSchematic = remote.resolveFile(schematicName);
-            if (remoteSchematic.exists()) {
-                remoteSchematic.delete();
-                if (remote.getChildren().length == 0) {
-                    remote.delete();
+            if (!onlyFinished) {
+                remote = fileManager.resolveFile(ftpURL.replaceFirst("finishedSchematics/",""), fileOptions);
+                remoteSchematic = remote.resolveFile(schematicName);
+                if (remoteSchematic.exists()) {
+                    remoteSchematic.delete();
+                    if (remote.getChildren().length == 0) {
+                        remote.delete();
+                    }
                 }
             }
 
@@ -109,8 +111,8 @@ public class FTPManager {
             remoteSchematic = remote.resolveFile(schematicName);
             if (remoteSchematic.exists()) {
                 remoteSchematic.delete();
-                if (remoteSchematic.getChildren().length == 0) {
-                    remoteSchematic.delete();
+                if (remote.getChildren().length == 0) {
+                    remote.delete();
                 }
             }
         }
